Move logic into actions
